+++ src/main/java/com/example/jpaboard/post/domain/Post.java
@@ -1,6 +1,8 @@
-package com.example.jpaboard.domain;
+package com.example.jpaboard.post.domain;
 
-import com.example.jpaboard.dto.PostRequest;
+import com.example.jpaboard.common.domain.BaseTimeEntity;
+import com.example.jpaboard.user.domain.User;
+import com.example.jpaboard.post.dto.PostRequest;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -10,7 +12,6 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.Lob;
 import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
