+++ src/main/java/com/example/jpaboard/post/infra/PostRepository.java
@@ -1,6 +1,6 @@
-package com.example.jpaboard.repository;
+package com.example.jpaboard.post.infra;
 
-import com.example.jpaboard.domain.Post;
+import com.example.jpaboard.post.domain.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
