--- src/main/java/com/example/jpaboard/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package com.example.jpaboard.repository;
-
-import com.example.jpaboard.domain.User;
-import java.util.Optional;
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-}
